Prepare current project for new-auth-api branch

+++ src/users/users.controller.ts
@@ -1,48 +1,66 @@
-import { Body, Controller, Delete, Get, Param, Patch, Post, Query } from '@nestjs/common';
+import { Body, Controller, Delete, ForbiddenException, Get, Param, Patch, Post, Query, Req, UseGuards } from '@nestjs/common';
 import { UsersService } from './users.service';
-
 import { CreateUserDto } from './dto/user.dto';
-
 import { ApiQuery } from '@nestjs/swagger';
+import { Jwtauthguard } from '../auth/jwt-auth.guard';
+import { ApiBearerAuth } from '@nestjs/swagger';
 
+
+
+@ApiBearerAuth()
 @Controller('users')
 export class UsersController {
-
     constructor(private readonly usersService: UsersService) {}
 
-    @ApiQuery({
-    name: 'role',
-    required: false,
-})
-
-    @Get() 
-    findAll(
-        @Query('role') role?: "admin" | "user" | "guest"
-    ) {
+    @ApiQuery({ name: 'role', required: false })
+    @UseGuards(Jwtauthguard)
+    @Get()
+    findAll(@Query('role') role?: "admin" | "user" | "guest") {
         return this.usersService.findAll(role);
     }
 
+    @UseGuards(Jwtauthguard)
     @Get(':id')
     findOne(@Param('id') id: string) {
         return this.usersService.findOne(Number(id));
     }
 
+
     @Post()
-    create(@Body() createUserDto: CreateUserDto) {
-    
-        return this.usersService.create(createUserDto);
+    async create(@Body() createUserDto: CreateUserDto, @Req() req?: Request) {
+    const requester = req?.['user']; 
+
+    if (createUserDto.role === 'admin') {
+      if (!requester || requester.role !== 'admin') {
+        throw new ForbiddenException('You cannot create admin users');
+      }
     }
 
-    @Patch(':id')
-    update(@Param('id') id: string, @Body() userUpdate: CreateUserDto) {
+    return this.usersService.create(createUserDto);
+  }
 
+    @UseGuards(Jwtauthguard)
+    @Patch(':id')
+    async update(@Param('id') id: string, @Body() userUpdate: CreateUserDto, @Req() req: Request) {
+        const requester = req['user'];
+        const targetUser = await this.usersService.findOne(Number(id));
+        if (requester.role !== 'admin' && targetUser && targetUser.role === 'admin') {
+            throw new ForbiddenException('You do not have permission to update this user');
+        }
         return this.usersService.update(Number(id), userUpdate);
     }
 
+    @UseGuards(Jwtauthguard)
     @Delete(':id')
-    delete(@Param('id') id: string) {
+    async delete(@Param('id') id: string, @Req() req: Request) {
+        const requester = req['user'];
+        const targetUser = await this.usersService.findOne(Number(id));
+
+        if (targetUser && requester.role !== 'admin' && targetUser.role === 'admin') {
+            throw new ForbiddenException('You do not have permission to delete this user');
+        }
+        
+        
         return this.usersService.delete(Number(id));
     }
-
-
 }
